Change nelemsPerGPU to type size_t

+++ tests/allgather_test.cu
@@ -47,7 +47,7 @@ static double getTime(void)
 
 __constant__ mscclppDevConn_t constDevConns[16];
 
-__device__ void allgather0(mscclppDevConn_t devConn, int rank, int world_size, int remoteRank, int nelemsPerGPU)
+__device__ void allgather0(mscclppDevConn_t devConn, int rank, int world_size, int remoteRank, size_t nelemsPerGPU)
 {
   // this allgather is really simple and implemented as an alltoall
 
@@ -63,7 +63,7 @@ __device__ void allgather0(mscclppDevConn_t devConn, int rank, int world_size, i
   devConn.wait();
 }
 
-__device__ void allgather1(mscclppDevConn_t devConn, int rank, int world_size, int remoteRank, int nelemsPerGPU)
+__device__ void allgather1(mscclppDevConn_t devConn, int rank, int world_size, int remoteRank, size_t nelemsPerGPU)
 {
   // this allgather algorithm works as follows:
   // Step 1: GPU rank i sends data to GPU rank (i+1) % world_size
@@ -82,7 +82,7 @@ __device__ void allgather1(mscclppDevConn_t devConn, int rank, int world_size, i
   devConn.wait();
 }
 
-__global__ void kernel(int rank, int world_size, int nelemsPerGPU, int kernel)
+__global__ void kernel(int rank, int world_size, size_t nelemsPerGPU, int kernel)
 {
   // only use a single thread from each warp
   if (threadIdx.x % 32 != 0)
@@ -119,14 +119,14 @@ void print_usage(const char* prog)
 #endif
 }
 
-void initializeAndAllocateAllGatherData(int rank, int world_size, size_t dataSize, int nelemsPerGPU, int** data_h,
+void initializeAndAllocateAllGatherData(int rank, int world_size, size_t dataSize, size_t nelemsPerGPU, int** data_h,
                                         int** data_d)
 {
   CUDACHECK(cudaMalloc(data_d, dataSize));
   CUDACHECK(cudaMemset(*data_d, 0, dataSize));
 
   *data_h = new int[nelemsPerGPU * world_size];
-  for (int i = 0; i < nelemsPerGPU * world_size; i++) {
+  for (size_t i = 0; i < nelemsPerGPU * world_size; i++) {
     int val = i + 1;
     if (i / nelemsPerGPU == rank) {
       (*data_h)[i] = val;
@@ -319,7 +319,7 @@ int main(int argc, const char* argv[])
   if (parsedArgs.find("datasize") != parsedArgs.end()) {
     dataSize = std::stoi(parsedArgs["datasize"]);
   }
-  int nelemsPerGPU = dataSize / sizeof(int) / world_size;
+  size_t nelemsPerGPU = dataSize / sizeof(int) / world_size;
 
   if (rank == 0)
     printf("Initializing data for allgather test\n");
@@ -343,7 +343,7 @@ int main(int argc, const char* argv[])
   CUDACHECK(cudaMemcpy(data_h, data_d, dataSize, cudaMemcpyDeviceToHost));
   CUDACHECK(cudaDeviceSynchronize());
 
-  for (int i = 0; i < nelemsPerGPU * world_size; i++) {
+  for (size_t i = 0; i < nelemsPerGPU * world_size; i++) {
     int val = i + 1;
     if (data_h[i] != val) {
       printf("oh uh! data_h[%d] (%d) != val (%d)\n", i, data_h[i], val);
